Fix alphanet with >4 nodes
Issue #817

The bug was that the main loop of ClientTask processes one even from
each channel at a time, and for the channel that receives partial BLS
signatures it meant that it drained it slower than it produced new
requests.

The workaround for now is to request signatures with a higher interval
(50 ms -> 500 ms).

Some additional changes:
- change log format to output nanoseconds
- cache your own BLS signature and don't recompute on each request
- make the requester of BLS signatures request from all authorities
instead of up to his own index